TASK: Refactor workspace selector to ts
